Generate a beveled 3D polygon

This function generates a beveled 3D polygon from a straight skeleton.

Usage

generate_beveled_polygon(
  skeleton,
  bevel_offsets = generate_bevel(),
  bevel_heights = NULL,
  set_max_height = FALSE,
  max_height = NA,
  vertical_offset = 0,
  base = TRUE,
  base_height = 0,
  raw_offsets = FALSE,
  raw_heights = FALSE,
  swap_yz = TRUE,
  progress = TRUE,
  double_sided = FALSE,
  sides = FALSE,
  return_skeleton_polygons = FALSE,
  scale_all_max = FALSE,
  material = material_list(),
  bevel_material = NA,
  verbose = FALSE
)

Arguments

skeleton

Default `NULL`. A straight skeleton generated from the `skeletonize` function.

bevel_offsets

Default `NULL`. The offset(s) of the bevel.

bevel_heights

Default is set to `bevel_offsets`. Numeric vector specifying the heights of the bevels. Must be of the same length as `bevel_offsets`.

set_max_height

Default `FALSE`. A logical flag that controls whether to set the max height of the roof based on the `max_height` argument.

max_height

Default `1`. The maximum height of the polygon.

vertical_offset

Default `0`. The vertical offset of the polygon.

base

Default `TRUE`. A logical flag that controls whether to generate the bottom of the polygon.

base_height

Default `NA`. Height of the base, defaulting to `min(bevel_heights) + vertical_offset` .

raw_offsets

Default `FALSE`. A logical flag indicating whether the `bevel_offsets` are already in raw format and do not need to be multiplied by the maximum time of the skeleton.

raw_heights

Default `FALSE`. A logical flag indicating whether the `bevel_heights` are already in raw format and do not need to be multiplied by the maximum time of the skeleton.

swap_yz

Default `TRUE`. A logical flag that controls whether to swap the y and z coordinates in the resulting mesh. If `TRUE`, the y and z coordinates will be swapped.

progress

Default `TRUE`. A logical flag to control whether a progress bar is displayed during roof generation.

double_sided

Default `FALSE`. A logical flag that controls whether the polygon should be double-sided.

sides

Default `FALSE`. A logical flag on whether to draw the sides. This will automatically be set to `TRUE` if `base = TRUE` and the `base_height` is less than `vertical_offset`.

return_skeleton_polygons

Default `FALSE`. A logical flag that controls whether to return the skeleton polygons along with the 3D mesh.

scale_all_max

Default `FALSE`. If passing in a list of multiple skeletons with polygons, whether to scale each polygon to the overall max height, or whether to scale each max height to the maximum internal distance in the polygon.

material

Default `material_list()`. Interface to set the color/appearance/material options for the resulting `ray_mesh` mesh.

bevel_material

Default `NA`, uses the material specified in `material`. Interface to set the color/appearance/material options for the resulting `ray_mesh` bevel mesh.

verbose

Default `FALSE`. A logical flag to control whether additional timing information should be displayed.

Value

A 3D mesh of the beveled polygon model.

Examples