Generate a 3D roof model

This function generates a 3D roof model from a straight skeleton.

generate_roof(
  skeleton,
  max_height = NA,
  vertical_offset = 0,
  base = FALSE,
  base_height = 0,
  angle = 45,
  sides = FALSE,
  double_sided = FALSE,
  scale_all_max = FALSE,
  swap_yz = TRUE,
  progress = TRUE,
  material = material_list(),
  roof_material = NA,
  verbose = FALSE
)

Arguments

skeleton

Default `NULL`. A straight skeleton generated from the `skeletonize` function.

max_height

Default `NA`. The maximum height of the roof.

vertical_offset

Default `0`. The vertical offset of the roof.

base

Default `TRUE`. A logical flag that controls whether to generate the bottom of the roof.

base_height

Default `vertical_offset`. Height of the base.

angle

Default `45`. Angle of the roof.

sides

Default `FALSE`. A logical flag on whether to draw the sides. This will automatically be set to `TRUE` if `base = TRUE` and the `base_height` is less than `vertical_offset`.

double_sided

Default `FALSE`. A logical flag that controls whether the polygon should be double-sided.

scale_all_max

Default `FALSE`. If passing in a list of multiple skeletons with polygons, whether to scale each polygon to the overall max height, or whether to scale each max height to the maximum internal distance in the polygon.

swap_yz

Default `TRUE`. A logical flag that controls whether to swap the y and z coordinates in the resulting mesh. If `TRUE`, the y and z coordinates will be swapped.

progress

Default `TRUE`. A logical flag to control whether a progress bar is displayed during roof generation.

material

Default `material_list()`. Interface to set the color/appearance/material options for the resulting `ray_mesh` mesh.

roof_material

Default `NA`, uses the material specified in `material`. Interface to set the color/appearance/material options for the resulting `ray_mesh` rooftop mesh.

verbose

Default `FALSE`. A logical flag to control whether additional timing information should be displayed.

Value

A 3D mesh of the roof model.

Examples